\name{exam_cheat_sheet}
\alias{exam_cheat_sheet}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
exam_cheat_sheet
}
\usage{
exam_cheat_sheet(week_person)
}
\description{
Find the teacher and your own answers from the past 3 weeks.
}
\examples{
exam_cheat_sheet(week1_teacher)
exam_cheat_sheet(week2_self)
}
